+++ src/full/Agda/Interaction/Options/Base.hs
@@ -691,6 +691,7 @@ defaultOptions = Options
   , optGenerateVimFile       = False
   , optIgnoreInterfaces      = False
   , optIgnoreAllInterfaces   = False
+  , optWriteInterfaces       = True
   , optPragmaOptions         = defaultPragmaOptions
   , optOnlyScopeChecking     = False
   , optTransliterate         = False
@@ -1141,6 +1142,9 @@ ignoreInterfacesFlag o = return $ o { optIgnoreInterfaces = True }
 ignoreAllInterfacesFlag :: Flag CommandLineOptions
 ignoreAllInterfacesFlag o = return $ o { optIgnoreAllInterfaces = True }
 
+noWriteInterfacesFlag :: Flag CommandLineOptions
+noWriteInterfacesFlag o = return $ o { optWriteInterfaces = False }
+
 traceImportsFlag :: Maybe String -> Flag CommandLineOptions
 traceImportsFlag arg o = do
   mode <- case arg of
@@ -1371,8 +1375,10 @@ standardOptions =
 
     , Option []     ["vim"] (NoArg vimFlag)
                     "generate Vim highlighting files"
-    , Option []     ["ignore-interfaces"] (NoArg ignoreInterfacesFlag)
+    , Option []     ["ignore-interfaces"] (NoArg ignoreInterfacesFlag)  -- TODO: rename to no-read-interfaces?
                     "ignore interface files (re-type check everything)"
+    , Option []     ["no-write-interfaces"] (NoArg noWriteInterfacesFlag)
+                    "don't write interface files to the filesystem"
     , Option ['i']  ["include-path"] (ReqArg includeFlag "DIR")
                     "look for imports in DIR"
     , Option ['l']  ["library"] (ReqArg libraryFlag "LIB")