+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Pragmas and options
 * New option `--print-options` to print a simple list of all options.
   This list can e.g. be used to implement bash completion.
 
+* New flag `--no-write-interfaces`
+
 Warnings
 --------